requirements.txt
Django>=3.2.6,<3.2.7
djangorestframework>=3.12.4,<3.12.5
psycopg2==2.9.1
flake8>=3.6.0,<3.7.0 